requirements/base.txt
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
Pillow==7.0.0
psycopg2_binary==2.8.4
pytz==2019.3
requests==2.23.0
six==1.14.0
celery==4.4.2
django-celery-beat==2.0.0
djangorestframework-simplejwt==4.4.0